lambda/kb-embedding/requirements.txt
boto3>=1.28.0
tiktoken>=0.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0
